+++ nodes/ccu-rpc-event.js
@@ -1,6 +1,20 @@
 const path = require('path');
 
 const statusHelper = require(path.join(__dirname, '/lib/status.js'));
+const {valueStatus} = require(path.join(__dirname, '/lib/valuestatus.js'));
+
+const FILTER_ATTRIBUTES = [
+    'rooms',
+    'functions',
+    'device',
+    'deviceName',
+    'deviceType',
+    'channel',
+    'channelName',
+    'channelType',
+    'channelIndex',
+    'datapoint',
+];
 
 module.exports = function (RED) {
     class CcuRpcEventNode {
@@ -21,20 +35,15 @@ module.exports = function (RED) {
                 iface: config.iface,
             };
 
+            // opt-in: only present when the option is enabled, so existing
+            // subscriptions keep receiving uncertain values (#96)
+            if (config.certainOnly) {
+                filter.uncertain = false;
+            }
+
             this.ccu.register(this);
 
-            [
-                'rooms',
-                'functions',
-                'device',
-                'deviceName',
-                'deviceType',
-                'channel',
-                'channelName',
-                'channelType',
-                'channelIndex',
-                'datapoint',
-            ].forEach((attr) => {
+            FILTER_ATTRIBUTES.forEach((attr) => {
                 if (!config[attr]) {
                     return;
                 }
@@ -45,8 +54,14 @@ module.exports = function (RED) {
                     filter[attr] = config[attr];
                 }
             });
+
+            // The node used to show nothing at all; show what came through
+            // last and how long ago (#52)
+            this.valueStatus = valueStatus(this);
+
             this.idSubscription = this.ccu.subscribe(filter, (message) => {
                 message.topic = this.ccu.topicReplace(config.topic, message);
+                this.valueStatus.set(message.topic + ' = ' + message.payload, message.ts);
                 this.send(message);
             });
             this.on('close', this._destructor);
@@ -54,6 +69,10 @@ module.exports = function (RED) {
 
         _destructor(done) {
             this.debug('ccu-rpc-event close');
+            if (this.valueStatus) {
+                this.valueStatus.stop();
+            }
+
             this.ccu.unsubscribe(this.idSubscription);
             this.ccu.deregister(this);
             done();